2019师大初赛

+++ Algorithm/Contest/2019-04-27-贵师大半决赛/01.cpp
@@ -1,10 +1,10 @@
-#include <bits/stdc++.h>
+#include <iostream>
 using namespace std;
 typedef long long ll;
 
 int main()
 {
-    int A, B;
+    ll A, B;
     cin >> A >> B;
     cout << A + B << endl;
     return 0;
